Add files via upload
# Kickstarting with Excel

## Overview of Project

### Purpose

The goal of this project was to create an analysis to determine what factors (if any) can help make the play successful. We are doing this by sorting and organizing data from various art forms so Louise can kickstart her crowd funding campaign to determine 


## Analysis and Challenges

### Analysis of Outcomes Based on Launch Date

My analysis for outcomes based on launch data would be that there are peak seasons for entertainment success. By season, I mean that success runs for longer than a month. First being summer months (when school is typically out for summer) and then back again for October and November during the fall. Also, February seems to be a peak month as well for successful. 


### Analysis of Outcomes Based on Goals

My analysis for outcomes based on goals would be that as the goal got bigger, the success rate went down which in turn means the percentage failed and the percentage canceled went up. I also looked at the totals of pledged funds and arrived at the conclusion that as the goal amount went higher the pledges we not as high. So that would mean with less money, the production was lacking in multiple areas such as set design, wardrobe, etc. 

### Challenges and Difficulties Encountered

My biggest challenge with this would have to be learning excel. To be honest, I’ve only known how to make things look pretty within excel and formulas scared me. I powered through and that leads me to believe the difficulty was me. Several times I had to stop and reel myself back because I was overwhelmed and/or thinking too much into it. Once I got out of my own head and I understood something on my own without any tips from google, I was on a roll.

## Results

- What are two conclusions you can draw about the Outcomes based on Launch Date?

Two conclusions that I arrived to are that theater seems like a risky venture as only 60% of the total productions have been successful and if you want to break out as a play write, the best time to do launch anything would be mid-summer as that is when the most success happens. 

- What can you conclude about the Outcomes based on Goals?

I would conclude that just because the goal is higher, does not mean you will be as successful as one would hope to be with more money. 

- What are some limitations of this dataset?

I do not feel there was any limitations on this dataset as I was able to break things down specifically to get tiny detailed needed/wanted. 

- What are some other possible tables and/or graphs that we could create?

I would say any graph would the word cluster in it. Examples would be Cluster Column or Cluster Bar specifically because they include each outcome for all the individual months while being separated by year. For me, that would have been helpful for comparing outcomes because I like having all the information, I’m comparing up at one once.